backend/api/requirements.txt
fastapi
uvicorn
pymongo
sqlalchemy
mariadb
Pillow
python-multipart